chore: Refactor OTEL bridge signals setup
refactor metrics setup

refactor logs setup
