+++ lib/otel/setup.js
@@ -6,18 +6,17 @@
 'use strict'
 
 const opentelemetry = require('@opentelemetry/api')
-const logsApi = require('@opentelemetry/api-logs')
-const { BasicTracerProvider } = require('@opentelemetry/sdk-trace-base')
 
-const NrSpanProcessor = require('./span-processor')
-const NrSampler = require('./sampler')
+const SetupLogs = require('./logs/index.js')
+const SetupMetrics = require('./metrics/index.js')
+const SetupTraces = require('./traces/index.js')
 const ContextManager = require('./context-manager')
 const TracePropagator = require('./trace-propagator')
 const defaultLogger = require('../logger').child({ component: 'opentelemetry-bridge' })
 const createOtelLogger = require('./logger')
 const interceptSpanKey = require('./span-key-interceptor')
-const bootstrapMetrics = require('./metrics/bootstrap-metrics')
-const bootstrapLogs = require('./logs/bootstrap-logs')
+
+const signals = []
 
 function setupOtel(agent, logger = defaultLogger) {
   if (agent.config.opentelemetry_bridge.enabled !== true) {
@@ -38,19 +37,22 @@ function setupOtel(agent, logger = defaultLogger) {
   opentelemetry.propagation.setGlobalPropagator(new TracePropagator(agent))
 
   if (agent.config.opentelemetry_bridge.traces.enabled === true) {
-    bootstrapTraces(agent)
+    const signal = new SetupTraces({ agent })
+    signals.push(signal)
   } else {
     logger.debug('`opentelemetry_bridge.traces` is not enabled, skipping')
   }
 
   if (agent.config.opentelemetry_bridge.metrics.enabled === true) {
-    bootstrapMetrics(agent)
+    const signal = new SetupMetrics({ agent })
+    signals.push(signal)
   } else {
     logger.debug('`opentelemetry_bridge.metrics` is not enabled, skipping')
   }
 
   if (agent.config.opentelemetry_bridge.logs.enabled === true) {
-    bootstrapLogs({ agent })
+    const signal = new SetupLogs({ agent })
+    signals.push(signal)
   } else {
     logger.debug('`opentelemetry_bridge.logs` is not enabled, skipping')
   }
@@ -60,30 +62,18 @@ function setupOtel(agent, logger = defaultLogger) {
     .incrementCallCount()
 }
 
-function bootstrapTraces(agent) {
-  opentelemetry.trace.setGlobalTracerProvider(new BasicTracerProvider({
-    sampler: new NrSampler(),
-    spanProcessors: [new NrSpanProcessor(agent)],
-    generalLimits: {
-      attributeValueLengthLimit: 4095
-    }
-  }))
-
-  agent.metrics
-    .getOrCreateMetric('Supportability/Nodejs/OpenTelemetryBridge/Traces')
-    .incrementCallCount()
-}
-
 function teardownOtel(agent) {
   if (agent?.config?.opentelemetry_bridge?.enabled !== true) {
     return
   }
 
-  opentelemetry.trace.disable()
+  for (const signal of signals) {
+    signal.teardown()
+  }
+
   opentelemetry.context.disable()
   opentelemetry.propagation.disable()
   opentelemetry.diag.disable()
-  logsApi.logs.disable()
 }
 
 module.exports = {
