+++ lib/otel/traces/attr-reconciler.js
@@ -5,8 +5,8 @@
 
 'use strict'
 
-const urltils = require('../util/urltils')
-const constants = require('./constants')
+const urltils = require('#agentlib/util/urltils.js')
+const constants = require('./constants.js')
 
 const hostKeys = [
   constants.ATTR_NET_HOST_NAME,
