+++ lib/otel/traces/exception-mapping.js
@@ -4,7 +4,7 @@
  */
 
 'use strict'
-const constants = require('./constants')
+const constants = require('./constants.js')
 const attrMappings = {
   msg: {
     attrs: [constants.EXCEPTION_MESSAGE, constants.EXCEPTION_TYPE]
