+++ lib/otel/traces/rules.js
@@ -6,7 +6,7 @@
 'use strict'
 
 const { SpanKind } = require('@opentelemetry/api')
-const srcJson = require('./transformation-rules')
+const srcJson = require('./transformation-rules.json')
 
 /**
  * Mapping rules are stored, and provided to the agent, as JSON. We call these
