+++ lib/otel/traces/segment-synthesis.js
@@ -4,8 +4,9 @@
  */
 
 'use strict'
-const { RulesEngine } = require('./rules')
-const defaultLogger = require('../logger').child({ component: 'segment-synthesizer' })
+
+const defaultLogger = require('#agentlib/logger.js').child({ component: 'segment-synthesizer' })
+const { RulesEngine } = require('./rules.js')
 const {
   createConsumerSegment,
   createDbSegment,
@@ -13,7 +14,7 @@ const {
   createInternalSegment,
   createProducerSegment,
   createServerSegment
-} = require('./segments')
+} = require('./segments/index.js')
 
 class SegmentSynthesizer {
   constructor(agent, { logger = defaultLogger } = {}) {
