+++ lib/otel/traces/segments/consumer.js
@@ -7,10 +7,10 @@
 
 module.exports = createConsumerSegment
 
-const Transaction = require('../../transaction/')
-const recorder = require('../../metrics/recorders/message-transaction')
-const { propagateTraceContext } = require('./utils')
-const createInternalSegment = require('./internal')
+const Transaction = require('#agentlib/transaction/index.js')
+const recorder = require('#agentlib/metrics/recorders/message-transaction.js')
+const { propagateTraceContext } = require('./utils.js')
+const createInternalSegment = require('./internal.js')
 
 function createConsumerSegment(agent, otelSpan, rule) {
   const activeTx = agent.tracer.getTransaction()
