+++ lib/otel/traces/segments/database.js
@@ -5,12 +5,12 @@
 
 'use strict'
 
-const parseSql = require('../../db/query-parsers/sql')
-const recordQueryMetrics = require('../../metrics/recorders/database')
-const recordOperationMetrics = require('../../metrics/recorders/database-operation')
-const ParsedStatement = require('../../db/parsed-statement')
-const metrics = require('../../metrics/names')
-const { transformTemplate } = require('../utils')
+const parseSql = require('#agentlib/db/query-parsers/sql.js')
+const recordQueryMetrics = require('#agentlib/metrics/recorders/database.js')
+const recordOperationMetrics = require('#agentlib/metrics/recorders/database-operation.js')
+const ParsedStatement = require('#agentlib/db/parsed-statement.js')
+const metrics = require('#agentlib/metrics/names.js')
+const { transformTemplate } = require('../utils.js')
 
 module.exports = function createDbSegment(agent, otelSpan, rule) {
   const context = agent.tracer.getContext()
