+++ lib/otel/traces/segments/http-external.js
@@ -5,10 +5,10 @@
 
 'use strict'
 
-const recordExternal = require('../../metrics/recorders/http_external')
-const urltils = require('../../util/urltils')
-const { transformTemplate } = require('../utils')
-const { UNKNOWN } = require('../constants')
+const recordExternal = require('#agentlib/metrics/recorders/http_external.js')
+const urltils = require('#agentlib/util/urltils.js')
+const { transformTemplate } = require('../utils.js')
+const { UNKNOWN } = require('../constants.js')
 
 function assignHost(otelSpan, segmentTransformation) {
   let host = UNKNOWN
