+++ lib/otel/traces/segments/internal.js
@@ -4,20 +4,16 @@
  */
 
 'use strict'
+const genericRecorder = require('#agentlib/metrics/recorders/generic.js')
 
-const genericRecorder = require('../../metrics/recorders/generic')
-const { transformTemplate } = require('../utils')
-
-module.exports = function createProducerSegment(agent, otelSpan, rule) {
+module.exports = function createInternalSegment(agent, otelSpan, rule) {
   const context = agent.tracer.getContext()
-  const segmentTransformation = rule.segmentTransformation
-  const name = transformTemplate(segmentTransformation?.name?.template, otelSpan?.attributes)
-
+  const name = otelSpan.name
   const segment = agent.tracer.createSegment({
     id: otelSpan?.spanContext()?.spanId,
     name,
-    recorder: genericRecorder,
     parent: context.segment,
+    recorder: genericRecorder,
     transaction: context.transaction
   })
   return { segment, transaction: context.transaction, rule }
