+++ lib/otel/traces/segments/producer.js
@@ -4,16 +4,20 @@
  */
 
 'use strict'
-const genericRecorder = require('../../metrics/recorders/generic')
 
-module.exports = function createInternalSegment(agent, otelSpan, rule) {
+const genericRecorder = require('#agentlib/metrics/recorders/generic.js')
+const { transformTemplate } = require('../utils.js')
+
+module.exports = function createProducerSegment(agent, otelSpan, rule) {
   const context = agent.tracer.getContext()
-  const name = otelSpan.name
+  const segmentTransformation = rule.segmentTransformation
+  const name = transformTemplate(segmentTransformation?.name?.template, otelSpan?.attributes)
+
   const segment = agent.tracer.createSegment({
     id: otelSpan?.spanContext()?.spanId,
     name,
-    parent: context.segment,
     recorder: genericRecorder,
+    parent: context.segment,
     transaction: context.transaction
   })
   return { segment, transaction: context.transaction, rule }
