+++ lib/otel/traces/segments/server.js
@@ -5,10 +5,10 @@
 
 'use strict'
 
-const Transaction = require('../../transaction')
-const httpRecorder = require('../../metrics/recorders/http')
-const { propagateTraceContext } = require('./utils')
-const createInternalSegment = require('./internal')
+const Transaction = require('#agentlib/transaction/index.js')
+const httpRecorder = require('#agentlib/metrics/recorders/http.js')
+const { propagateTraceContext } = require('./utils.js')
+const createInternalSegment = require('./internal.js')
 
 module.exports = function createServerSegment(agent, otelSpan, rule) {
   const activeTx = agent.tracer.getTransaction()
