+++ lib/otel/traces/segments/utils.js
@@ -5,7 +5,7 @@
 
 'use strict'
 
-const Traceparent = require('../../w3c/traceparent')
+const Traceparent = require('#agentlib/w3c/traceparent.js')
 
 /**
  * Accepts trace context payload if span has a parent. It will use the
