+++ lib/otel/traces/span-processor.js
@@ -7,19 +7,19 @@
 
 const { hrTimeToMilliseconds } = require('@opentelemetry/core')
 const { SpanKind } = require('@opentelemetry/api')
-const urltils = require('../util/urltils')
+const urltils = require('#agentlib/util/urltils.js')
 
-const AttributeReconciler = require('./attr-reconciler')
-const SegmentSynthesizer = require('./segment-synthesis')
-const { otelSynthesis } = require('../symbols')
-const { DESTINATIONS } = require('../config/attribute-filter')
-const { assignToTarget, buildRuleMappings, extractAttributeValue, processRegex, transformTemplate } = require('./utils')
-const defaultLogger = require('../logger').child({ component: 'span-processor' })
+const AttributeReconciler = require('./attr-reconciler.js')
+const SegmentSynthesizer = require('./segment-synthesis.js')
+const { otelSynthesis } = require('#agentlib/symbols.js')
+const { DESTINATIONS } = require('#agentlib/config/attribute-filter.js')
+const { assignToTarget, buildRuleMappings, extractAttributeValue, processRegex, transformTemplate } = require('./utils.js')
+const defaultLogger = require('#agentlib/logger.js').child({ component: 'span-processor' })
 
 const {
   SPAN_STATUS_CODE
 } = require('./constants')
-const exceptionAttr = require('./exception-mapping')
+const exceptionAttr = require('./exception-mapping.js')
 
 module.exports = class NrSpanProcessor {
   #reconciler
