+++ lib/otel/traces/transformation-rules.json
