+++ lib/otel/traces/utils.js
@@ -4,7 +4,8 @@
  */
 
 'use strict'
-const { DESTINATIONS } = require('../config/attribute-filter')
+
+const { DESTINATIONS } = require('#agentlib/config/attribute-filter.js')
 
 /**
  * Transforms a template string by replacing placeholders with values from the data object.
