+++ test/unit/lib/otel/exception-mapping.test.js
@@ -4,12 +4,12 @@
  */
 
 'use strict'
-const exceptionAttr = require('#agentlib/otel/exception-mapping.js')
+const exceptionAttr = require('#agentlib/otel/traces/exception-mapping.js')
 const test = require('node:test')
 const assert = require('node:assert')
 const {
   EXCEPTION_TYPE
-} = require('#agentlib/otel/constants.js')
+} = require('#agentlib/otel/traces/constants.js')
 
 test('msg', () => {
   const span = {
