+++ test/unit/lib/otel/fixtures/producer.js
@@ -12,7 +12,7 @@ const {
   ATTR_MESSAGING_DESTINATION_NAME,
   ATTR_MESSAGING_SYSTEM,
   ATTR_MESSAGING_OPERATION,
-} = require('#agentlib/otel/constants.js')
+} = require('#agentlib/otel/traces/constants.js')
 
 function createProducerSpan({ tracer, name = 'test-span' }) {
   const span = createSpan({ name, kind: SpanKind.PRODUCER, tracer })
